# Use EscapeMetricName from statsd-exporter for sanitize