--- a/go.mod
+++ b/go.mod
@@ -5,6 +5,7 @@ require (
 	github.com/golang/groupcache v0.0.0-20191027212112-611e8accdfc9 // indirect
 	github.com/prometheus/client_golang v1.2.1
 	github.com/prometheus/procfs v0.0.6 // indirect
+	github.com/prometheus/statsd_exporter v0.15.0
 	go.opencensus.io v0.22.2
 	golang.org/x/sys v0.0.0-20191113165036-4c7a9d0fe056 // indirect
 )