--- a/sanitize.go
+++ b/sanitize.go
@@ -15,8 +15,7 @@
 package prometheus
 
 import (
-	"strings"
-	"unicode"
+	"github.com/prometheus/statsd_exporter/pkg/mapper"
 )
 
 const labelKeySizeLimit = 100
@@ -30,21 +29,10 @@ func sanitize(s string) string {
 	if len(s) > labelKeySizeLimit {
 		s = s[:labelKeySizeLimit]
 	}
-	s = strings.Map(sanitizeRune, s)
-	if unicode.IsDigit(rune(s[0])) {
-		s = "key_" + s
-	}
+
+	s = mapper.EscapeMetricName(s)
 	if s[0] == '_' {
 		s = "key" + s
 	}
 	return s
-}
-
-// converts anything that is not a letter or digit to an underscore
-func sanitizeRune(r rune) rune {
-	if unicode.IsLetter(r) || unicode.IsDigit(r) {
-		return r
-	}
-	// Everything else turns into an underscore
-	return '_'
 }