--- a/sanitize_test.go
+++ b/sanitize_test.go
@@ -65,3 +65,31 @@ func TestSanitize(t *testing.T) {
 		})
 	}
 }
+
+func BenchmarkSanitize(b *testing.B) {
+	inputs := []string{
+		"grpc.io/client/sent_bytes_per_rpc",
+		"grpc.io/client/received_bytes_per_rpc",
+		"grpc.io/client/roundtrip_latency",
+		"grpc.io/client/completed_rpcs",
+		"grpc.io/client/sent_messages_per_rpc",
+		"grpc.io/client/received_messages_per_rpc",
+		"grpc.io/client/server_latency",
+		"grpc.io/server/received_bytes_per_rpc",
+		"grpc.io/server/sent_bytes_per_rpc",
+		"grpc.io/server/server_latency",
+		"grpc.io/server/completed_rpcs",
+		"grpc.io/server/received_messages_per_rpc",
+		"grpc.io/server/sent_messages_per_rpc",
+		"grpc_server_method",
+		"grpc_server_status",
+		"grpc_client_method",
+		"grpc_client_status",
+	}
+
+	for n := 0; n < b.N; n++ {
+		for _, in := range inputs {
+			_ = sanitize(in)
+		}
+	}
+}