See https://github.com/prometheus/statsd_exporter/blob/master/pkg/mapper/escape.go

EscapeMetricName is highly optimized for this use case.

Metric names in benchmark are from https://godoc.org/go.opencensus.io/plugin/ocgrpc

Benchmark comparison:

```
$ benchcmp /tmp/a.txt /tmp/b.txt
benchmark                old ns/op     new ns/op     delta
BenchmarkSanitize-12     2861          1304          -54.42%

benchmark                old allocs     new allocs     delta
BenchmarkSanitize-12     13             13             +0.00%

benchmark                old bytes     new bytes     delta
BenchmarkSanitize-12     624           544           -12.82%
```